setup storybook for docs
setup storybook for docs

sort components alphabetically

always show code in storybook doc
