+++ .storybook/main.js
@@ -1,4 +1,4 @@
 module.exports = {
-	stories: [`../src/**/*stories.tsx`],
-	addons: ['@storybook/addon-backgrounds', '@storybook/addon-viewport'],
+	stories: ['../src/**/*.stories.mdx', '../src/**/*stories.@(js|jsx|ts|tsx)'],
+	addons: ['@storybook/addon-links', '@storybook/addon-essentials'],
 };
