--- .storybook/manager-head.html
@@ -1,5 +1,0 @@
-<style>
-	#storybook-preview-iframe {
-		border: 0 !important;
-	}
-</style>
