+++ .storybook/preview/backgrounds.ts
@@ -1,0 +1,31 @@
+import {
+	background,
+	brandBackground,
+	brandAltBackground,
+} from '@guardian/src-foundations';
+
+export const backgrounds = {
+	default: 'background.primary',
+	values: [
+		{
+			name: 'background.primary',
+			value: background.primary,
+		},
+		{
+			name: 'background.secondary',
+			value: background.secondary,
+		},
+		{
+			name: 'background.inverse',
+			value: background.inverse,
+		},
+		{
+			name: 'brandBackground.primary',
+			value: brandBackground.primary,
+		},
+		{
+			name: 'brandAltBackground.primary',
+			value: brandAltBackground.primary,
+		},
+	],
+};
