+++ README.md
@@ -2,6 +2,8 @@
 
 > A component library for the Guardian's [Source Design System](https://theguardian.design).
 
+[![Storybook](https://cdn.jsdelivr.net/gh/storybookjs/brand@main/badge/badge-storybook.svg)](https://guardian.github.io/source)
+
 ## Components ⚛️
 
 See [the Source storybook](https://guardian.github.io/source) for examples of available components.
