+++ package.json
@@ -65,9 +65,10 @@
     "@emotion/babel-preset-css-prop": "^11.0.0",
     "@emotion/react": "^11.1.2",
     "@guardian/prettier": "^0.6.0",
-    "@storybook/addon-backgrounds": "^6.3.4",
-    "@storybook/addon-viewport": "^6.3.4",
-    "@storybook/react": "^6.3.4",
+    "@storybook/addon-actions": "^6.3.5",
+    "@storybook/addon-essentials": "^6.3.5",
+    "@storybook/addon-links": "^6.3.5",
+    "@storybook/react": "^6.3.5",
     "@types/jest": "^26.0.20",
     "@typescript-eslint/eslint-plugin": "^4.0.0",
     "@typescript-eslint/parser": "^3.10.1",
